Move Zebra.Serial -> Zebra.Binary
This is so I can do a Zebra.JSON format for import/export
