+++ zebra-core/src/Zebra/Binary.hs
@@ -1,0 +1,8 @@
+{-# LANGUAGE NoImplicitPrelude #-}
+module Zebra.Binary (
+    module X
+  ) where
+
+import           Zebra.Binary.Array as X
+import           Zebra.Binary.Block as X
+import           Zebra.Binary.Header as X
