--- zebra-core/src/Zebra/Serial.hs
@@ -1,8 +1,0 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-module Zebra.Serial (
-    module X
-  ) where
-
-import           Zebra.Serial.Array as X
-import           Zebra.Serial.Block as X
-import           Zebra.Serial.Header as X
