+++ zebra-core/test/test.hs
@@ -1,5 +1,10 @@
 import           Disorder.Core.Main
 
+import qualified Test.Zebra.Binary.Array
+import qualified Test.Zebra.Binary.Block
+import qualified Test.Zebra.Binary.File
+import qualified Test.Zebra.Binary.Header
+import qualified Test.Zebra.Binary.Table
 import qualified Test.Zebra.Data.Block
 import qualified Test.Zebra.Data.Core
 import qualified Test.Zebra.Foreign.Block
@@ -8,17 +13,17 @@ import qualified Test.Zebra.Foreign.Merge
 import qualified Test.Zebra.Foreign.Table
 import qualified Test.Zebra.Merge.Entity
 import qualified Test.Zebra.Schema
-import qualified Test.Zebra.Serial.Array
-import qualified Test.Zebra.Serial.Block
-import qualified Test.Zebra.Serial.File
-import qualified Test.Zebra.Serial.Header
-import qualified Test.Zebra.Serial.Table
 import qualified Test.Zebra.Table
 
 main :: IO ()
 main =
   disorderMain [
-      Test.Zebra.Data.Block.tests
+      Test.Zebra.Binary.Array.tests
+    , Test.Zebra.Binary.Block.tests
+    , Test.Zebra.Binary.File.tests
+    , Test.Zebra.Binary.Header.tests
+    , Test.Zebra.Binary.Table.tests
+    , Test.Zebra.Data.Block.tests
     , Test.Zebra.Data.Core.tests
     , Test.Zebra.Foreign.Block.tests
     , Test.Zebra.Foreign.Entity.tests
@@ -26,10 +31,5 @@ main =
     , Test.Zebra.Foreign.Table.tests
     , Test.Zebra.Merge.Entity.tests
     , Test.Zebra.Schema.tests
-    , Test.Zebra.Serial.Array.tests
-    , Test.Zebra.Serial.Block.tests
-    , Test.Zebra.Serial.File.tests
-    , Test.Zebra.Serial.Header.tests
-    , Test.Zebra.Serial.Table.tests
     , Test.Zebra.Table.tests
     ]
